Move Controller test to live next to implementation
