+++ package.json
@@ -70,6 +70,7 @@
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "^7.1.0",
     "@types/jest": "^24.0.0",
+    "@types/mock-raf": "^1.0.2",
     "@types/react": "16.8.2",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "24.1.0",
